Removed callbacks from jsx
